requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.16
botocore==1.34.16
celery==5.3.6
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.7
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
ffmpeg-python==0.2.0
future==0.18.3
greenlet==3.0.3
gunicorn==21.2.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.4
msgpack==1.0.7
packaging==23.2
prompt-toolkit==3.0.43
psycopg==3.1.13
psycopg-binary==3.1.13
psycopg-pool==3.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
s3transfer==0.10.0
service-identity==23.1.0
six==1.16.0
sqids==0.4.1
SQLAlchemy==1.4.51
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
zope.interface==6.1
